requirements.txt
discord.py
Pillow
praw
ffmpeg
sympy 
moviepy
gtts
youtube_dl
PyNaCl
python-dotenv